Separete functions that are used outside the options page
